+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All Notable changes to `flipboxdigital\craft-sortable-associations` will be documented in this file
 
+## Unreleased
+### Changed
+- `SortableAssociations` service no longer uses a constant to identify the record alias.
+- `SortableFields` service no longer uses a constant to identify the record alias.
+
 ## 1.0.0-rc.1 - 2018-03-23
 ### Changed
 - `SortableAssociation` record must implement an association service.